feat(search): support EF_RUNTIME for HNSW vector search (#7696)
* feat(search): support EF_RUNTIME for HNSW vector search

* fix: review comments

* fix: review comments

+++ src/core/search/ast_expr.cc
@@ -63,7 +63,7 @@ AstTagsNode::AstTagsNode(AstExpr&& l, TagValue tag) {
 }
 
 AstKnnNode::AstKnnNode(uint32_t limit, std::string_view field, OwnedFtVector vec,
-                       std::string_view score_alias, std::optional<size_t> ef_runtime)
+                       std::string_view score_alias, std::optional<uint32_t> ef_runtime)
     : filter{nullptr},
       limit{limit},
       field{field.substr(1)},
@@ -95,7 +95,7 @@ bool AstKnnNode::HasPreFilter() const {
 }  // namespace dfly::search
 
 namespace std {
-ostream& operator<<(ostream& os, optional<size_t> o) {
+ostream& operator<<(ostream& os, optional<uint32_t> o) {
   return os;
 }
